Even more refactoring on imports
