+++ client/src/components/hooks/index.js
@@ -1,0 +1,5 @@
+import useGames from './useGames';
+import useKey from './useKey';
+import useLocalStorageState from './useLocalStorageState';
+
+export { useGames, useKey, useLocalStorageState };
