+++ client/src/components/hooks/useKey.jsx
@@ -1,6 +1,6 @@
 import { useEffect } from 'react';
 
-export function useKey(key, action) {
+export default function useKey(key, action) {
   useEffect(
     function () {
       function callback(e) {
