+++ client/src/components/hooks/useLocalStorageState.jsx
@@ -1,5 +1,5 @@
 import { useState, useEffect } from 'react';
-export function useLocalStorageState(initialState, key) {
+export default function useLocalStorageState(initialState, key) {
   const [value, setValue] = useState(function () {
     const storedValue = localStorage.getItem(key);
     return storedValue ? JSON.parse(storedValue) : initialState;
